Perform repo initialization via array of structs
The method of initializing the repo via a map[string]**sqlx.NamedStmt
usually works well, until you hit the case where two identical SQL
statements should be registered somewhere (this probably happens more in
tests than it does in real life, but it can plausibly happen).

Making the initialization loop operate on an array of structures
sidesteps this problem quite nicely: With this change, every SQL
statement can get registered on a NamedStmt field, no matter what its
shape or how similar it is to other statements on the repository.